Extract the mongodb conf file loading into a separate util class, in preparation of changes needed to ensure the correct order of operations for setting the authentication username/password for the global auth flag.
